simplifia-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
google-cloud-aiplatform==1.38.0
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.14.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0